removes authentication
