+++ app/index.tsx
@@ -1,16 +1,6 @@
 import { Redirect } from "expo-router";
-import React from "react";
-import { useAuth } from "./hooks/useAuth";
 
 export default function Index() {
-  const { isAuthenticated, loading } = useAuth();
-
-
-  if (loading) {
-    return null; // Root layout will show loading
-  }
-
-  // Simple one-time redirect based on auth status
-  
-  return isAuthenticated ? <Redirect href="/(tabs)" /> : <Redirect href="/(auth)/login" />;
+  // Redirect directly to tabs (no authentication required)
+  return <Redirect href="/(tabs)" />;
 }
